dev: added MenuButton template
