+++ src/UIkit/MenuButton/MenuButton.module.css
@@ -1,0 +1,6 @@
+.root {
+  max-width: 1920px;
+  padding: 0;
+  margin: 0 auto;
+  text-align: center;
+}
