+++ src/UIkit/MenuButton/MenuButton.test.tsx
@@ -1,0 +1,7 @@
+import { describe, expect, test } from 'vitest'
+
+describe('XComponent', () => {
+  test('should add two numbers', () => {
+    expect(1 + 1).toBe(2)
+  })
+})
