+++ src/UIkit/MenuButton/index.ts
@@ -1,0 +1,3 @@
+import MenuButton from './MenuButton'
+
+export default MenuButton
